requirements.txt
smbus
struct
time
signal